requirements.txt
numpy==1.22.1
Pillow==9.0.1